+++ contrib/autopilot/autopilot.py
@@ -43,8 +43,8 @@ from contrib.autopilot.tasks import TEXT_UTF8_RAW_BITEXT_REVERSE, TEXT_REQUIRES_
 from contrib.autopilot.tasks import TEXT_UTF8_TOKENIZED
 from contrib.autopilot.tasks import RAW_FILES
 from contrib.autopilot.tasks import Task, TASKS
-from contrib.autopilot.models import MODELS, MODEL_NONE, MODEL_TEST_ARGS
-from contrib.autopilot.models import DECODE_ARGS, DECODE_STANDARD
+from contrib.autopilot.models import MODELS, MODEL_NONE, MODEL_GNMT, MODEL_TEST_ARGS
+from contrib.autopilot.models import DECODE_ARGS, DECODE_STANDARD, DECODE_GNMT
 from contrib.autopilot import third_party
 
 
@@ -325,7 +325,7 @@ def renew_step_dir(step_dir: str):
     os.makedirs(step_dir)
 
 
-def call_sockeye_train(args: List[str],
+def call_sockeye_train(model: str,
                        bpe_dir: str,
                        model_dir: str,
                        log_fname: str,
@@ -336,7 +336,7 @@ def call_sockeye_train(args: List[str],
     partial training or skip training if model is already finished.  Record
     command for future use.
 
-    :param args: Command line arguments for sockeye.train.
+    :param model: Type of translation model to train.
     :param bpe_dir: Directory of BPE-encoded input data.
     :param model_dir: Model output directory.
     :param log_fname: Location to write log file.
@@ -351,7 +351,7 @@ def call_sockeye_train(args: List[str],
               "--validation-target={}".format(os.path.join(bpe_dir, PREFIX_DEV + SUFFIX_TRG_GZ)),
               "--output={}".format(model_dir)]
     # Assemble command
-    command = [sys.executable, "-m", "sockeye.train"] + fnames + args
+    command = [sys.executable, "-m", "sockeye.train"] + fnames + MODELS[model]
     # Request GPUs or specify CPU
     if num_gpus > 0:
         command.append("--device-ids=-{}".format(num_gpus))
@@ -359,7 +359,7 @@ def call_sockeye_train(args: List[str],
         command.append("--use-cpu")
     # Test mode trains a smaller model for a small number of steps
     if test_mode:
-        command += MODEL_TEST_ARGS
+        command += MODEL_TEST_ARGS[model]
     command_fname = os.path.join(model_dir, FILE_COMMAND.format("sockeye.train"))
     # Run unless training already finished
     if not os.path.exists(command_fname):
@@ -721,6 +721,8 @@ def run_steps(args: argparse.Namespace):
 
     logging.info("=== Train translation model ===")
     logging.info("Model: %s", args.model)
+    if args.model == MODEL_GNMT:
+        logging.info("NOTE: This is an 8 layer LSTM model similar (but not exactly identical) to the 'GNMT' architecture.")
     step_dir_model = os.path.join(dir_task, DIR_PREFIX_MODEL + args.model)
     complete_fname = os.path.join(step_dir_model, FILE_COMPLETE)
     if os.path.exists(complete_fname):
@@ -729,7 +731,7 @@ def run_steps(args: argparse.Namespace):
         log_fname = os.path.join(args.workspace,
                                  DIR_LOGS,
                                  "sockeye.{{}}.{}.{}.{}.log".format(task_name, args.model, os.getpid()))
-        call_sockeye_train(MODELS[args.model],
+        call_sockeye_train(args.model,
                            step_dir_bpe,
                            step_dir_model,
                            log_fname.format("train"),