+++ CHANGELOG.md
@@ -10,10 +10,38 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [1.18.45]
+### Added
+- Added an 8 layer LSTM model similar (but not exactly identical) to the 'GNMT' architecture to autopilot.
+
+## [1.18.44]
+### Fixed
+- Fixed an issue with `--max-num-epochs` causing training to stop before the update/batch that actually completes the epoch was made.
+
+## [1.18.43]
+### Added
+- `<s>` now supported as the first token in a multi-word negative constraint
+  (e.g., `<s> I think` to prevent a sentence from starting with `I think`)
+### Fixed
+- Bugfix in resetting the state of a multiple-word negative constraint
+
+## [1.18.42]
+### Changed
+- Simplified gluon blocks for length calculation
+
+## [1.18.41]
+### Changed
+- Require numpy 1.14 or later to avoid MKL conflicts between numpy as mxnet-mkl.
+
+## [1.18.40]
+### Fixed
+- Fixed bad check for existence of negative constraints.
+- Resolved conflict for phrases that are both positive and negative constraints.
+- Fixed softmax temperature at inference time.
 
 ## [1.18.39]
-### Added 
-- Image Captioning now supports constrained decoding. 
+### Added
+- Image Captioning now supports constrained decoding.
 - Image Captioning: zero padding of features now allows input features of different shape for each image.
 
 ## [1.18.38]
@@ -55,7 +83,7 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
 
 ## [1.18.30]
 ### Fixed
-- Preserving max output length for each sentence to allow having identical translations for both with and without batching. 
+- Preserving max output length for each sentence to allow having identical translations for both with and without batching.
 
 ## [1.18.29]
 ### Changed